Union final de la interfaz y el lector con las diferentes funciones del proyecto :white_check_mark:
