+++ nbproject/private/private.properties
@@ -17,7 +17,7 @@ user.properties.file=C:\\Users\\santi\\AppData\\Roaming\\NetBeans\\19\\build.pro
 <<<<<<< HEAD
 user.properties.file=C:\\Users\\PC\\AppData\\Roaming\\NetBeans\\20\\build.properties
 =======
-user.properties.file=C:\\Users\\PC\\AppData\\Roaming\\NetBeans\\21\\build.properties
+user.properties.file=C:\\Users\\santi\\AppData\\Roaming\\NetBeans\\19\\build.properties
 >>>>>>> 376f84cb61210bcd1ea502fcb95e727c22b3eee1
 >>>>>>> a05d45269855439778aea537f20808e7215abfb8
 >>>>>>> 3b039f213daf62d3fd89dfb8f3f8ac97d5282690
