+++ nbproject/private/private.xml
@@ -3,20 +3,18 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/interfaz/t3.png</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/ArbolReserva.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/main.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/Registro.java</file>
-            <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/interfaz/t5.png</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/ListaHistorial.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/Cliente.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/ArbolHabitacion.java</file>
-            <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/LectorCsv/LectorCSV.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/Reservacion.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/ListaReserva.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/FuncionesHotel.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/interfaz/Home.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/Habitacion.java</file>
+            <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/LectorXlsx/LectorXlsx.java</file>
             <file>file:/C:/Users/santi/OneDrive/Desktop/Proyecto2_EDD_Sec3_Feliciani-Fermoso-Oliveros/EDD_Proyecto2-Feliciani-Fermoso-Oliveros/src/Estructuras/Historial.java</file>
         </group>
     </open-files>
