+++ nbproject/project.properties
@@ -35,12 +35,12 @@ dist.jlink.dir=${dist.dir}/jlink
 dist.jlink.output=${dist.jlink.dir}/Java_Proyecto2
 endorsed.classpath=
 excludes=
-file.reference.commons-collections4-4.4.jar=../../commons-collections4-4.4.jar
-file.reference.commons-compress-1.20.jar=../../commons-compress-1.20.jar
-file.reference.poi-5.0.0.jar=../../poi-5.0.0.jar
-file.reference.poi-ooxml-5.0.0.jar=../../poi-ooxml-5.0.0.jar
-file.reference.poi-ooxml-full-5.0.0.jar=../../poi-ooxml-full-5.0.0.jar
-file.reference.xmlbeans-4.0.0.jar=../../xmlbeans-4.0.0.jar
+file.reference.commons-collections4-4.4.jar=LibreriasLector/commons-collections4-4.4.jar
+file.reference.commons-compress-1.20.jar=LibreriasLector/commons-compress-1.20.jar
+file.reference.poi-5.0.0.jar=LibreriasLector/poi-5.0.0.jar
+file.reference.poi-ooxml-5.0.0.jar=LibreriasLector/poi-ooxml-5.0.0.jar
+file.reference.poi-ooxml-full-5.0.0.jar=LibreriasLector/poi-ooxml-full-5.0.0.jar
+file.reference.xmlbeans-4.0.0.jar=LibreriasLector/xmlbeans-4.0.0.jar
 includes=**
 jar.compress=false
 javac.classpath=\
