+++ src/Estructuras/Habitacion.java
@@ -5,8 +5,10 @@
 package Estructuras;
 
 /**
- *
- * @author santi
+ * @Description Main
+ * @author Santiago
+ * @version 23/03/24
+ * @retorna Habitacion
  */
 public class Habitacion {
     private int num_hab;
