+++ src/Estructuras/Reservacion.java
@@ -5,8 +5,10 @@
 package Estructuras;
 
 /**
- *
- * @author santi
+ * @Description Main
+ * @author Santiago
+ * @version 23/03/24
+ * @retorna Reservacion
  */
 public class Reservacion {
