+++ src/main.java
@@ -4,7 +4,7 @@ import Estructuras.ArbolReserva;
 import Estructuras.FuncionesHotel;
 import Estructuras.Habitacion;
 import Estructuras.Registro;
-import LectorCsv.LectorCSV;
+import LectorXlsx.LectorXlsx;
 import interfaz.Home;
 
 /*
@@ -12,9 +12,11 @@ import interfaz.Home;
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Main.java to edit this template
  */
 /**
- *
- * @author santi
+ * @Description Main
+ * @author Santiago
+ * @version 23/03/24
  */
+
 public class main {
 
     /**
@@ -23,51 +25,7 @@ public class main {
     public static void main(String[] args) {
         // TODO code application logic here
 
-//        Registro registroprueba = new Registro(7);
-//        registroprueba.InsertarHash("a", "b", 1);
-//        registroprueba.InsertarHash("c", "d", 2);
-//        registroprueba.InsertarHash("e", "f", 3);
-//        System.out.println(registroprueba.ImprimirHash());
-//        registroprueba.EliminarHash("nombre", "fvefvfev");
-//        System.out.println(registroprueba.ImprimirHash());
-////        System.out.println(registroprueba.BuscarHash("erge", "fveffdvdvfvfev").getPrimer_nombre());
-//
-//        System.out.println("Hola");
-//        System.out.println("ayudapapadio");
-//        System.out.println("prueba github");
-//        System.out.println("prueba github2");
-//
-//        ArbolReserva arbol = new ArbolReserva();
-//        arbol.IntertarReser(arbol.getRaiz(), 30, "san", "feli", "g@gmail.com", "male", "2", "04242140509", "2/3/04", "2/3/03");
-//        arbol.IntertarReser(arbol.getRaiz(), 8, "sa", "fel", "g@gmail.co", "mal", "3", "0424214050", "2/3/4", "/3/03");
-//        arbol.IntertarReser(arbol.getRaiz(), 78, "s", "fe", "g@gmail.c", "ma", "4", "042421405", "2//04", "23/03");
-//        arbol.IntertarReser(arbol.getRaiz(), 7, "s", "fe", "g@gmail.c", "ma", "4", "042421405", "2//04", "23/03");
-//        arbol.IntertarReser(arbol.getRaiz(), 900, "s", "fe", "g@gmail.c", "ma", "4", "042421405", "2//04", "23/03");
-//        arbol.IntertarReser(arbol.getRaiz(), 888, "s", "fe", "g@gmail.c", "ma", "4", "042421405", "2//04", "23/03");
-//        arbol.IntertarReser(arbol.getRaiz(), 18, "s", "fe", "g@gmail.c", "ma", "4", "042421405", "2//04", "23/03");
-//
-//        ArbolHabitacion arbol2 = new ArbolHabitacion();
-//        arbol2.IntertarHab(arbol2.getRaiz(), 1, "doueble", "1");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 2, "3", "1");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 3, "2", "1");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 4, "4", "2");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 5, "3", "2");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 6, "4", "2");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 7, "4", "3");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 8, "3", "3");
-//        arbol2.IntertarHab(arbol2.getRaiz(), 9, "2", "3");
-//
-//        Habitacion h = arbol2.BuscarHab(arbol2.getRaiz(), 1);
-//        h.getHistorial().InsertarLista(10, "nom", "ape", "mail", "gen", "lle", 1);
-//        h.getHistorial().InsertarLista(20, "nom1", "ape1", "mail1", "gen1", "lle1", 1);
-//        h.getHistorial().InsertarLista(30, "nom2", "ape2", "mail2", "gen2", "lle2", 1);
-//        h.getHistorial().InsertarLista(40, "nom3", "ape3", "mail3", "gen3", "lle3", 1);
-//
-////        System.setProperty("log4j2.loggerContextFactory","com.simba.neo4j.shaded.apache.logging.log4j.simple.SimpleLoggerContextFactory");
-//        System.out.println("hola");
-//        System.out.println("al dia");
-        
-        LectorCSV lector = new LectorCSV();
+        LectorXlsx lector = new LectorXlsx();
         lector.LeerCSV();
         FuncionesHotel fh = new FuncionesHotel(lector.a, lector.x, lector.y);
